NP-657 Move general functions to utils
