+++ src/pages/publication/submission_tab/ErrorSummary.tsx
@@ -1,26 +1,22 @@
 import React, { FC } from 'react';
-import { FormikProps, useFormikContext, FormikErrors } from 'formik';
+import { FormikProps, useFormikContext } from 'formik';
 import styled from 'styled-components';
 import { FormikPublication } from '../../../types/publication.types';
 import Heading from '../../../components/Heading';
 import NormalText from '../../../components/NormalText';
 import { useTranslation } from 'react-i18next';
 import Card from '../../../components/Card';
+import { flattenFormikErrors } from '../../../utils/formik-helpers';
 
 const StyledCard = styled(Card)`
   border: 3px solid ${({ theme }) => theme.palette.danger.main};
   background-color: ${({ theme }) => theme.palette.danger.light};
 `;
 
-interface FormikError {
-  fieldName: string;
-  errorMessage: string;
-}
-
 const ErrorSummary: FC = () => {
   const { t } = useTranslation('publication');
   const { errors }: FormikProps<FormikPublication> = useFormikContext();
-  const flattenedErrors = flattenErrors(errors);
+  const flattenedErrors = flattenFormikErrors(errors);
 
   return flattenedErrors.length > 0 ? (
     <StyledCard>
@@ -33,18 +29,6 @@ const ErrorSummary: FC = () => {
       ))}
     </StyledCard>
   ) : null;
-};
-
-// Convert all errors from nested object to flat array
-const flattenErrors = (validationErrors: FormikErrors<any>): FormikError[] => {
-  return Object.entries(validationErrors)
-    .map(([fieldName, errorMessage]) => {
-      if (typeof errorMessage === 'object' && errorMessage !== null) {
-        return flattenErrors(errorMessage as FormikErrors<any>);
-      }
-      return { fieldName, errorMessage };
-    })
-    .flat();
 };
 
 export default ErrorSummary;
